+++ backend/canisters/group/api/src/main.rs
@@ -78,7 +78,6 @@ fn main() {
     generate_ts_method!(group, video_call_participants);
 
     generate_ts_method!(group, accept_p2p_swap);
-    generate_ts_method!(group, add_bot);
     generate_ts_method!(group, add_reaction);
     generate_ts_method!(group, block_user);
     generate_ts_method!(group, cancel_invites);
@@ -97,7 +96,6 @@ fn main() {
     generate_ts_method!(group, register_poll_vote);
     generate_ts_method!(group, register_proposal_vote);
     generate_ts_method!(group, register_proposal_vote_v2);
-    generate_ts_method!(group, remove_bot);
     generate_ts_method!(group, remove_participant);
     generate_ts_method!(group, remove_reaction);
     generate_ts_method!(group, report_message);