+++ backend/canisters/group/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Disallow forwarding prize messages and governance proposal messages ([#7260](https://github.com/open-chat-labs/open-chat/pull/7260))
 - Move storage of group bots up a level ([#7265](https://github.com/open-chat-labs/open-chat/pull/7265))
 - Introduce new `Integer` bot parameter type ([#7290](https://github.com/open-chat-labs/open-chat/pull/7290))
+- Sync bot installation with UserIndex ([#7291](https://github.com/open-chat-labs/open-chat/pull/7291))
 
 ### Fixed